pyproject.toml
[tool.poetry]
name = "scoutf"
version = "0.1.0"
description = ""
authors = ["codewithdpk <153810687+linkinlab@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-hcl2 = "^5.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
